Add vending machine circuit boards and construction
This adds the ability to construct vending machines, reducing the need for
admin intervention in the case that a crucial machine gets destroyed early
in the round. Some machines don't have restock crates, but the majority of
important ones do.

If a vending machine is constructed, it starts with an empty inventory; if
destroyed, it spews some inventory as is currently the case, and then sets
down a machineframe that can be rebuilt into the machine (depending on the
level of damage done). When rebuilt, the lost inventory remains lost.

As a side effect, this PR adds empty variants of vending machines, so when
appropriate mappers can place those into their maps.